Warn when --show-all is used with no files listed
At least one contributor thought that --show-all would show all owners
for the current CL, so they used it with no files specified and got
confused by the lack of output. I was fooled to. The help message makes
it clear but it seems worth detecting the no-files case and printing a
helpful message, especially since most (all?) of the other git cl owners
commands apply to the current CL.

Bug: 1071967
Change-Id: I77bd2e30a3b3a6eb67cd40b93f5544b20facacb8