Fix extra spaces in git cl completion
The commands come from the output of git cl -h, which then goes through
an sed filter, matching on the ANSI escape sequence for green and reset.
This ends up having extra spaces, since the output aligns the
subcommands based on the longest subcommand. For shorter commands like
archive, it becomes "archive   ". So, tweak the sed regex to only
capture the part without the whitespaces.

Change-Id: I80a84f142520478547895f85def68331575fdd72