Fix some Python3 bugs in package_from_installed.py
The package_from_installed.py script doesn't run under Python 3. This
change gets it closer by using universal_newlines=True so that
subprocess.check_output returns a string instead of bytes, and
by explicitly converting dict.items() to a list.

This also changes the default Windows 10 SDK version to something more
recent - the current version that we build with.

To test:

  python3 package_from_installed.py -d 2019

Unfortunately this still fails, but it gets much farther. The last
failure is that this command:

  json.dump({'env': collections.OrderedDict(list(env.items()) + list(env_x86.items()))},

Fails with TypeError: a bytes-like object is required, not 'str'

Similarly, this fails:

    print(VS_VERSION, file=version)

TypeError: a bytes-like object is required, not 'str'
Change-Id: I2e3fc4d464f2914ba5ae36f0a78570ea7e5d4f03