Update patch set 1
Patch Set 1:

I want to make sure depot tools is getting updated on our bots if this CL is landed. I'm not sure if there's a mechanism already. I have a feeling it might just be this auto updating. Does anyone know? I can dig into it or ask around if needed.

Patch-set: 1
CC: Gerrit User 1148599 <1148599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>