Update patch set 6
Patch Set 6:

> > > Patch Set 4: -Auto-Submit
> > > 
> > > Osato-san, could you confirm this improve time no-op build using autoninja?
> > 
> > Wow, it quickened almost double!
> > ToT:
> > $ tools\win\timeit autoninja.bat -C out\Debug base                                                                                         
> > "C:\src\build\depot_tools\ninja.exe" -C out\Debug base -j 1000
> > ninja: Entering directory `out\Debug'
> > ninja: no work to do.
> > 9.50 seconds elapsed time
> > 
> > This patch:
> > $ tools\win\timeit autoninja.bat -C out\Debug base                                                                                         
> > "C:\src\build\depot_tools\ninja.exe" -C out\Debug base -j 1000
> > ninja: Entering directory `out\Debug'
> > ninja: no work to do.
> > 5.84 seconds elapsed time
> 
> So this shows almost the same performance with direct ninja invocation?

I see ninja.exe taking ~2.8 seconds on my laptop, so still much faster. I suspect that the 9.50/5.84 change was randomness, since I see no improvement from this change. Or, the improvement might be the tendency of second runs of ninja to be faster due to caching effects.

Patching in https://crrev.com/c/2235916 makes it easier to see the cost of ninja_upload in isolation, which makes it easier to see whether it's speed has improved.

I think that information about speed improvements should go in the commit message instead of just in the discussion comments - that is my personal preference anyway.

Patch-set: 6