Update patch set 1
Patch Set 1:

As mentioned on the thread, this is the cpplint.py from github, so we probably shouldn't edit it. Our depot tools script presubmit_canned_checks.py specifies the filter we want to use, so we could remove it explicitly there with a -runtime/references.

I noticed since sending that email that we have a BLACKLIST_LINT_FILTERS (should rename this) which drops build/c++11, and that's probably the place to add this one too.

Patch-set: 1
CC: Gerrit User 1000887 <1000887@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>