Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> As mentioned on the thread, this is the cpplint.py from github, so we probably shouldn't edit it.

Given that that cpplint.py is supposed to enforce the Google style guide and the Google style guide has changed globally, doesn't it make sense to change this?

> Our depot tools script presubmit_canned_checks.py specifies the filter we want to use, so we could remove it explicitly there with a -runtime/references.

I think this is more prone to subdirectories intentionally or unintentionally not using this filter, though.  I was surprised by the subtle behavior interactions with this and per-directory lint checks while trying to add more linting in ui/views/, for example, and I think now ui/views/ ignores this value (and IIRC cc/ does too?).  It's more predictable and reliable if we can change cpplint.py.

> I noticed since sending that email that we have a BLACKLIST_LINT_FILTERS (should rename this) which drops build/c++11, and that's probably the place to add this one too.

I dunno if we should explicitly blacklist this, since in theory the rule works OK if for some reason a specific directory needs to use it, it's just that people shouldn't default to using it.

Patch-set: 1
Label: Commit-Queue=0