gclient: Silently ignore SyntaxErrors when eval'ing gclient-to-GN vars.
To see what this will fix, consider this innocuous change:
https://chromium-review.googlesource.com/c/chromium/src/+/2267854

That adds a '-' to a gclient var that gets passed down to GN. But due
to how gclient handles these kinds of vars, it considers the value to
be an arithmetic expression, when all we really want is a string
literal.

This CL fixes that by only trying to evaluate the var, and if it fails,
passing it along unmodified.

Bug: 937821
Change-Id: Ia4cef70d305a686575a49930371db2ba8ba0cf8b