mac: Add an arm64 slice to ninja-mac.
To build the arm ninja binary, I downloaded a zip with the ninja 1.8.2
code, and ran `./configure --bootstrap` on an arm mac, with Xcode 12 beta
installed, followed by `strip ninja`. That gave me a working ninja 1.8.2
arm binary. I then copied it to my intel mac and ran

    lipo -create -output ninja-mac ninja-mac ~/Downloads/ninja

to make a fat binary.

Bug: 1103315,762354
Change-Id: I2c4bea53a01d7a6135f9cd5554e3123dba960664