Handle BUG_PREFIX correctly when populating Bug: lines.
Right now git_cl.py inserts an extra colon. Remove it, and sprinkle in
some asserts to make sure the BUG_PREFIX ends with a colon already.

This is happening because _get_bug_line_values() assumes the input is
the project name, but all the callers pass in the bug prefix. Fix up the
unit tests as well, so the inputs are all prefixes.

Bug: 1111955
Change-Id: I3db199435087abb17b9600fc208b11d2d8e6a615