[Backward Compatible] Timing and RDB for Lower-level checks
Allows presubmit_support to perform Timing and RDB reporting for each of
the lower-level "CheckXYZ" functions, provided that the variable
NEW_PRESUBMIT_CHECKS_ENABLED is set to true in the PRESUBMIT.py
file that we are running. I am using the variable to avoid the breakage
caused by the similar previous CL which was not backwards compatible.
In addition, asks for "-realm" with ResultDB so it can maintain the
security boundary.

Bug: 1106943
Change-Id: Ia49e8884e2653bee7de7ed7d254452d55dd5c617