Support specifying buildtools in .gclient
Currently CHROMIUM_BUILDTOOLS_PATH can be used to override the
buildtools. But since it's a environment variable, it'll affect all
checkouts on the machine.

This CL adds an option to specify "buildtools" directly in the .gclient
file, so that we can have per-checkout buildtools override.

Here's an example of .gclient file with buildtools override:

solutions = [
  { "name"        : 'foo/bar',
    "url"         : 'https://foosource.com/bar',
    "deps_file"   : 'DEPS',
    "managed"     : False,
    "buildtools"  : 'foo/buildtools',
  },
]

Bug: b/162095628
Change-Id: Ia997b0a9687496ab05289ffe67946d8c4da47aaa