Pass large -j value to ninja when use_rbe = true
The goma-on-borg backend is turning down
(go/goma-borg-turndown-faq). While chromium can continue using
the goma client for build acceleration, the backend is swapped
out to instead use Remote Build Execution, a gcloud service.

Other downstream projects (e.g. Nest) switched out the goma
frontend as well as the backend to RBE. These projects use a
|use_rbe| gn flag to enable build acceleration through RBE.

This change allows autoninja to recognize this flag and treat it
as if it was using goma for the purposes of passing an
appropriate -j flag.

Bug: None
Test: run on an out/ directory with use_rbe enabled

Change-Id: I52d5771e92e9163234811b154fd36816639bf1b8