Xcode 12b3 (and later) emits arm64e binaries; try that architecture.
In Xcode 12 beta 3, pointer authentication was turned on for ARM64
binaries, yielding arm64e architecture. Attempting to run with only
arm64 specified does not allow an arm64e slice to run. Prefer arm64e
if available, but allow an arm64 slice if that's the only option.
(If provided multiple architectures, arch uses the first one that can
be used.)

Bug: 1103236
Change-Id: Id3498816a2495a449a16b3b6a66b0e776b70f4e0