Fix goma regex detection and add optional nice launching
The detection of goma in autoninja does not work for the use_goma flag
when it does not simply appear at the beginning of the line.

Here we add slightly improved detection for the goma flag.

Also adding linux nice process launch feature to improve the user
experience when running a large build. This feature is disabled by
default like other autoninja features.



R=brucedawson@chromium.org

Change-Id: I8a04e7480c5e8d37d57fb71143c6066d46abb004