Reduce git cl split oversplitting.
Without this change, git cl split will create a separate CL for every
file that has per-file owners rules. This is almost always the wrong
thing to do. With this change, it will at least aggregate all such files
to their parent directory. This may under-split, but is perhaps nearer
the mark.

Change-Id: Iea7997b36d053713fa00f1261ea032e09d9c2818