Use HEAD instead of hardcoded branch
Repositories may have different default branches. This ensures to fetch
remote HEAD branch which is default branch. Only downside to this
approach is that actual branch (e.g. origin/main) reference won't be
updated. Since we don't need it for this particular case, it should be
fine to use HEAD.

R=apolito@google.com, ehmaldonaldo@chromium.org

Bug: 1155492
Change-Id: Ib2e6e50901a0a10ed647acea3aeb517b83567702