Add reproxy setup/teardown to autoninja.
Adds reproxy setup and teardown to autoninja.  Since reproxy is intended
to run for a single build (unlike the goma proxy), having setup and
teardown happen as part of autoninja makes management of the proxy
execution easier.

To use this as it is currently implemented, set the RBE_BIN_DIR and
RBE_CFG_DIR environment variables to point to the reclient binaries and
reclient configuration files.  Note the reproxy.cfg file is NOT included
in this change, and at the time of this CL you need to provide one
yourself.

Bug: 1149386
Change-Id: I23601cc9b13193ac617ffc5963b9d443f6840d33