Check whether goma is running when it is enabled
One of the mistakes one can make when running ninja is having goma
enabled (use_goma=true in args.gn) but not having goma running. This can
lead to ~1,000 failed compile steps, which is messy.

This change teaches autoninja.py to check whether goma is running. If
not then it tells autoninja.bat to just print a warning message. The
check costs roughly 150 ms which seems reasonable.

If build acceleration is requested through use_rbe then no checking for
whether the service is running is done. That could be added.

autoninja.py could auto-start goma but that is error prone and has
limited additional value.

Bug: 868590, b/174673874
Change-Id: Ie773e574878471e5136b9b82d52f86af3d848318