Add previous_got_revision and previous_revision_cp to output
If the bot has previously checked out the repo, output which revision
and commit position were checked out.

ensure_checkout() would now take in a `output_prev_checkout_info` kwarg
to determine whether that info should be printed out or not.
This change would make it so that every ensure_checkout call will
always attempt to look for the previous head (git rev-parse HEAD).
Please let me know if doing this call every time will cause issues, so
I can modify it so that it only does that call when the kwarg is passed
in.

Change-Id: Ib8a6f152a593446ce4082477bc9ada44f1fd1fc3