Require LF line endings
Chromium standard is for lines to end with line-feed ('\n') characters.
On Windows it is easy to generate files with carriage-return line-feed
('\r\n') line endings. Among other problems this causes the CheckLicense
presubmit to fail in an inscrutable way. That is, CheckLicense says that
a source file lacks a valid license, but manual inspection shows nothing
wrong.

This change adds a CR check in front of the CheckLicense check.

This is supposed to be checked by CheckForWindowsLineEndings in
src\PRESUBMIT.py. Testing has shown that this presubmit has multiple
bugs. Fixing them will resolve this. Abandoning.

Bug: 801033
Change-Id: I567881fe3f5bf9d4fc212c0f4b6d2fafd064180f