Require LF line endings
Chromium standard is for lines to end with line-feed ('\n') characters.
On Windows it is easy to generate files with carriage-return line-feed
('\r\n') line endings. Among other problems this causes the CheckLicense
presubmit to fail in an inscrutable way.

This change adds a CR check in front of the CheckLicense check. Let's
see how it goes! It may be necessary to only do this check if
CheckLicense fails.

Bug: 801033
Change-Id: I567881fe3f5bf9d4fc212c0f4b6d2fafd064180f