win toolchain: Prepare downloader for windows sdk dir switch
crrev.com/c/2655836 tries to move the Windows SDK from
"win_sdk" to "Windows Kits/10".

get_toolchain_if_necessary.py (in depot_tools) saves the path to the SDK to
third_party/depot_tools/win_toolchain/data.json which then gets copied
by a script in the chromium repo to build/win_toolchain.json.
For the SDK move to work, chromium's pinned depot_tools
must write the new SDK path when rolling in the new toolchain package.
So the plan is:

1. Land this change, which can handle both path styles
2. Wait for depot_tools in chromium to update
3. Then roll to a win toolchain package with the new layout

In a few years, when we no longer need the old layout,
we can remove this detection code again and assume the new layout.

Bug: 1173176
Change-Id: Iaefc5c16685d3dbfff87a3e50a7b20b457366e44