GLOBAL_APPROVERS prototype
This adds support for a //GLOBAL_APPROVERS file. The file acts as a top-
level OWNERS file, but it is only used for files or paths that are
specified in the CL description. For example

GA=chrome/browser
GA=components/printing

These lines would let anyone in GLOBAL_APPROVERS have their LGTM apply
as an approved OWNERS check for files under the specified directories.

This allows both the CL author and the reviewer to be know what is
being approved, and code archaeologists to see or audit as well.

** This protootype does not fully work **

TODO: There are some tests that do not pass, that should, and the
BatchListOwners needs some better python to compile.

Change-Id: I86a22cc4c783771ba0c6a9a7370933e6f435bf54