Don't make depot_tools pylintrc override local pylint configs.
In the current pylint handling, the only way to use an alternative
config is to explicitly pass it with the "--rcfile" flag, which is
confusingly inconsistent with normal pylint config file resolution.
This change allows local pylintrc files or PYLINTRC env var to override
the default depot_tools config, which is often desirable to enforce
different style guidelines for different projects.

Change-Id: Ib725b15bb639dc9c7cb9009fd3b504124e0c1f2a