Fix ninjalog_uploader crash when running opt-in or opt-out.
This CL removes the binary mode when opening the ninjalog.cfg file.

It looks like the json serializer operates with strings and it does
not like the binary mode of the file object argument.

Before this change, this is what happens:

√ ~/chromium/src % ../depot_tools/ninjalog_uploader_wrapper.py opt-out
Traceback (most recent call last):
  File "../depot_tools/ninjalog_uploader_wrapper.py", line 125, in <module>
    sys.exit(main())
  File "../depot_tools/ninjalog_uploader_wrapper.py", line 89, in main
    SaveConfig(config)
  File "../depot_tools/ninjalog_uploader_wrapper.py", line 41, in SaveConfig
    json.dump(config, f)
  File "/Library/Frameworks/Python.framework/Versions/3.6/lib/python3.6/json/__init__.py", line 180, in dump
    fp.write(chunk)
TypeError: a bytes-like object is required, not 'str'

Bug: 1181998
Change-Id: If18cc2220271e53be0a49abe97c85121870c3146