Make autoninja.py handle git bash better
If autoninja.py detects that goma is not running when it should be then
it prints 'cmd /c exit 1' as the command to be run. This ensures that
errorlevel will be set so that the failure will be detected.
Unfortunately this doesn't work when running autoninja from git bash.

This change adds quotes so that this works from git bash (invoking the
shell script) as well as from cmd.exe (invoking the batch file).

Bug: 868590
Change-Id: I482f22830f9bd4f7b70c51de9647a70d946ec145