Make autoninja.py work again w/ Python3 and MSYS/MinGW bash.
We switched autoninja.py to python3 in https://crrev.com/c/2752772
but that assumes `python3` is in $PATH (which is likely true on
Mac and Linux but less likely to be true on Windows, especially
in MSYS/MinGW/git bash shells.

It's safer to use the `python3` bundled into depot_tools in
`$depot_tools/python-bin` and more hermetic to boot.

Bug: 1187609
Change-Id: I65efd1a2b87135ac4e25ba377309c3374cfda085